+++ components/traders_comp/feedback/FeedbackForm.jsx
@@ -1,14 +1,54 @@
 "use client";
 
-import React from 'react';
+import React, { useState } from 'react';
+import Box from '@mui/material/Box';
+import Modal from '@mui/material/Modal';
+import Fade from '@mui/material/Fade';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
+import FeedbackForm from './FeedbackForm'; // Import your feedback form component
 
-
-const FeedbackForm = () => {
-    return (
-        <div>
-            <h1>Feedback Form</h1>
-        </div>
-    );
+const style = {
+  position: 'absolute',
+  top: '50%',
+  left: '50%',
+  transform: 'translate(-50%, -50%)',
+  width: 400,
+  bgcolor: 'background.paper',
+  border: '2px solid #000',
+  boxShadow: 24,
+  p: 4,
 };
 
-export default FeedbackForm;
+export default function TransitionsModal() {
+  const [open, setOpen] = useState(false);
+
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
+
+  return (
+    <div>
+      <Button onClick={handleOpen}>Feedback</Button> {/* Changed button text to "Feedback" */}
+      <Modal
+        aria-labelledby="transition-modal-title"
+        aria-describedby="transition-modal-description"
+        open={open}
+        onClose={handleClose}
+        closeAfterTransition
+        BackdropComponent={Box} // Changed slots to BackdropComponent
+        BackdropProps={{
+          timeout: 500,
+        }}
+      >
+        <Fade in={open}>
+          <Box sx={style}>
+            <Typography id="transition-modal-title" variant="h6" component="h2">
+              Feedback Form {/* Changed modal title */}
+            </Typography>
+            <FeedbackForm /> {/* Render the FeedbackForm component */}
+          </Box>
+        </Fade>
+      </Modal>
+    </div>
+  );
+}
